+++ packages/coordinator/src/screens/AuthScreens/AuthScreens.module.scss
@@ -1,14 +1,26 @@
-.screen {
-  width: 100%;
+.resetPasswordbackButton {
+  position: fixed;
+  top: 20px;
+  right: 20px;
+
+  display: inline-flex;
+  align-items: center;
+
+  font-size: 14px;
 }
 
-.loginScreenLogoContainer {
+.screen {
+  width: 100%;
+  padding-top: 50px;
+}
+
+.authScreenLogoContainer {
   display: flex;
   flex-direction: column;
   align-items: center;
 }
 
-img.loginScreenLogoImage {
+img.authScreenLogoImage {
   width: 120px;
 }
